// src/pages/inventory/transactions/sales/transaction-types.ts
export interface TransactionData {
  master: TransactionMaster;
  masterValidations?: TransactionValidationsData;
  details: TransactionDetail[];
  invAccTransactions: InvAccTransaction[];
  attachments: any[]
}
// ...
export interface TransactionDetail {
  invTransactionDetailID?: number;
  slNo: string;
  pCode: string;
  mrp: number;
  barCode: string;
  productBatchID: number;
  product: string;
  productID: number;
  brand: string;
  brandID: number;
  qty: number;
  qtyTag?: number;
  free: number;
  multiFactor: number;
  unit: string;
  unitID: number;
  unitDecimalPoint: number;
  unitPrice: number;
  unitPriceTag: number;
  gross: number;
  discPerc: number;
  discount: number;
  netValue: number;
  total: number;
  stock: number;
  manualBarcode: string;
  stockDetails: string;
  margin: number;
  salesPrice: number;
  lpr: number;//LastPurchaseRate
  lpc: number;//lastPurchaseCost
  stickerQty: number;
  profit: number;
  size: string;
  vatPerc: number;
  vatAmount: number;
  cst: number;


  cstPerc: number;
  cost: number;
  batchNo: string;
  mfdDate: string; // or Date
  expDate: string; // or Date
  expDays: number;
  bd: string;
  btnPrintBarcode: string;
  barcodePrinted: boolean;
  batchCreated: boolean;
  actionCol: boolean;
  productDescription: string;
  serial: string;
  minSalePrice: number;
  additionalExpense: number;
  unitPriceFC: number;
  colour: string;
  warranty: string;
  nosQty: number;
  totalAddExpense: number;
  grossConvert: number;
  grossFC: number;
  unitID2: number;
  unit2Qty: number;
  unit2SalesRate: number;
  unit2MRP: number;
  unit2MBarcode: string;
  unit2StickerQty: number;
  unitID3: number;
  unit3Qty: number;
  unit3SalesRate: number;
  unit3MRP: number;
  unit3MBarcode: string;
  unit3StickerQty: number;
  tagQty: number;
  barcodeTagPrinted: boolean;
  barcodeUnit2Printed: boolean;
  barcodeUnit3Printed: boolean;
  location: string;
  grTransDetailsID: number;
  grTransDetailsIDTag?: number;
  arabicName: string;
  supplierReferenceProductCode: string;
  poTransDetailsID: number;
  poTransDetailsIDTag: number;
  ratePlusTax: number;
  warehouseID: number;
  warehouseName: string;
  sortOrder: number;
  profitPercentage: number;
  schemeDiscount: number;
  moreDetails: TransactionDetailsMore;
  memoEditor: string;
  rowNumber: number;
  actualSalesPrice: number;
  unit2: string;
  unit3: string;
  btnPrintBarcodeStd: string;
  isValid?: boolean;
  hsnCode: string;
  details2?: TransactionDetails2;
  gRTransDetailID: number;
  pOTransDetailID: number;
  pO_PITransDetailIDs: number;
  pO_PITransDetailQtys: number;
}
// ...
export interface UserConfig {
  maxWidth?: any;
  gridMaxWidth?: any;
  gridHeight?: any;
  keepNarrationForJV?: boolean;
  clearDetailsAfterSaveAccounts?: boolean;
  mnuShowConfirmationForEditOnAccounts?: boolean;
  maximizeBillwiseScreenInitially?: boolean;
  alignment?: "left" | "center" | "right";
  presetCostenterId?: number;
  counterAssignedCashLedgerId?: number;
  outerPageBg?: string;
  innerPageBg?: string;
  inputBoxStyle?: inputBox;
  isExpanded?: boolean;
  useBarcode?: boolean;
  resizeGrid?: boolean;
  showProductInfoPopup?: boolean;
  showPurchaserOnly?: boolean;
  useSupplierProductCode?: boolean;
  presetWarehouseId?: number;
  counterWiseWarehouseId?: number;
  enableItemCodeSearchInNameColumn?: boolean;
  holdSameCode?: boolean;
  printPreview?: boolean;
  duplicationMessage?: boolean;
  setDefaultQuantity?: boolean;
  useInSearch?: boolean;
  useCodeSearch?: boolean;
  gridFontSize?: number;
  gridIsBold?: boolean;
  gridRowHeight?: number;
  gridBorderColor?: string;
  gridHeaderBg?: string;
  footerPosition?: "bottom" | "right";
  gridHeaderFontColor?: string;
  gridBorderRadius?: number;
  barCodePrev?: boolean;
  footerBg?: string;
  showColumnBorder?: boolean;
  activeRowBg?: string;
  useNewFooter?: boolean;
  themeName?: string;
  gridHeaderRowHeight?: number;
  gridFooterBg?: string;
  gridFooterFontColor?: string;
  enableVoucherPrefix?: boolean;
}

export type FormElementsState = {
  [key in keyof typeof initialFormElements]: FormElementState;
};

export interface LoadData {
  formType?: string;
  vPrefix?: string;
  vNumber?: string;
}

export interface SummaryItems {
  qty: number;
  gross: number;
  netValue: number;
  discount: number;
  vatAmount: number;
  cst: number;
  total: number;
  profit: number;
  grossFc: number;
  totalAddExpense: number;
  nosQty: number;
  pCode: string;
  barCode: string;
  margin: number;
  salesPrice: string;
  cgst: number;
  sgst: number;
  igst: number;
  cessAmt: number;
  additionalCess: number;
}
// ...
export interface TransactionFormState {
  store: any;
  formCode: string;
  userRightsFormCode: string;
  isEntryControl: boolean;
  isEdit: boolean;
  isRowEdit: boolean;
  ledgerDataLoading: boolean;
  saving: boolean;
  savingCompleted?: boolean;
  ledgerBalanceLoading: boolean;
  ledgerBalance: number;
  ledgerData: any;
  groupName: any;
  dtLedgerCodes: any[];
  summary: SummaryItems;
  showPartySelection: boolean;
  showSaveDialog: boolean;
  customerType: string;
  isInvoker: boolean;
  title: string;
  partyId?: string;
  prev: string;
  rowProcessing: boolean;
  transactionProcessing: boolean;
  transactionLoading: boolean;
  unlocking: boolean;
  transaction: TransactionData;
  transactionType: string;
  total: number;
  printOnSave: boolean;
  printPreview: boolean;
  printCheque: boolean;
  amountInWords: string;
  template?: any;
  templates?: [];
  templatesData?: TemplateState<TransactionDetail>[];
  userConfig?: UserConfig;
  formElements: FormElementsState;
  openUnsavedPrompt?: boolean;
  foreignCurrency: boolean;
  enableDebitAccount: boolean;
  inSearch: boolean;
  enableTaxNumber: boolean;
  tmpVoucherNo?: number;
  dummyCode?: any;
  dummyProducts?: boolean;
  remarks?: string;
  prevTransactionDate?: string;
  lb?: boolean;
  autoCalculation?: boolean;
  priceCategory: string;
  netTotal: number;
  netAmount: number;
  loadData: LoadData;
  isProductSummaryOpen: boolean;
  isPartyWiseSummaryOpen: boolean;
  isFormStateDetailOpen?: boolean;
  selectedRow?: TransactionDetail;
  gridColumns: ColumnModel[];
  isPostedTransaction: boolean;
  isInv: boolean;
  summaryConfig: SummaryConfig<TransactionDetail>[];
  showQuantityFactors: { visible: boolean; rowIndex: number, qtyDesc:string };
  showPcode: boolean;
  batchEntryData: { visible: boolean; data: string; rowIndex: number };
  serialNoEntryData: { visible: boolean; data: string; rowIndex: number };
  batchSelectionData: string;
  popupSearchSelectionData: string;
  quantityFactorData: string;
  currentCell?: CurrentCell;
  batchesUnits?: UnitByBatchDetailsDto[];
  productInfo: boolean;
  ShowProductBatchUnitDetails: boolean;
  showProductInformation?: { show: boolean; index: number };
  barcodeTemplate?: any;
  barcodeData?: any;
  barcodePrevOpen?: boolean;
  ledgerDetails: boolean;
  loading?: loadingResult;
  dataWarranty: [];
  dataBrands: [];
  headerMorePop: boolean;
  showGridTheme: boolean;
  // stockUpdate: boolean;
  batchGridShowKey?: number;
  orderStatus?: string;
  selectedTheme?: any
  currentTheme?: any
  themeChangeCountdown?: number
  isInitialLedger?: boolean;
  memoEditor: { visible: boolean; data: string; rowIndex: number };
  headerMenuOpen: boolean;
  documentModal: boolean;
  pendingOrdListMasterIDs?: string;
  pendingOrdListBranchIDs?: string;
  gridMenuOpen: boolean;
  billwiseData?: BillwiseData[];
  billwiseDrCr?: string;
  billwiseDetails?: string;
  showbillwise?: boolean;
  ledgerBillWiseLoading?: boolean;
}
interface loadingResult {
  isLoading: boolean;
  text: string;
}
export interface CurrentCell {
  reCenterRow?: boolean
  column: string;
  rowIndex: number;
  data: TransactionDetail;
}
export type PartialTransactionFormFields = {
  [K in keyof TransactionFormState]?: TransactionFormState[K];
};
export interface ColumnModel {
  dataField?: string;
  decimalPoint?: number;
  formStateOptionKey?: string;
  cssClass?: string;
  caption?: string;
  dataType?: "string" | "number" | "cb" | "date" | "boolean" | any;
  allowEditing?: boolean;
  alignment?: "center" | "left" | "right";
  format?: string;
  visible?: boolean;
  isLocked?: boolean;
  width?: number;
  readOnly?: boolean;
  idField?: string; //for cb
  field?: { valueKey: string; labelKey: string };
  cellRender?: (
    cellElement: any,
    cellInfo: any,
    filter?: any,
    pdfCell?: any
  ) => any;
}
export interface PrintTransProps {
  masterAccount: string;
  details: TransactionDetail[];
  voucherType: string;
  formType: string;
  vrPrefix: string;
  vrNumber: string;
  isPrintPreview: boolean;
}

export interface VoucherElementProps {
  formState: TransactionFormState;
  dispatch: Dispatch<AnyAction>;
  handleKeyDown?: (e: any, field: string) => void;
  t: any;
}
// ...
export interface UnitByBatchDetailsDto {
  unitCode?: string;
  productBatchID?: number; // C# long -> TS number
  value?: number;
  label?: string;
  multiFactor?: string;
  unitDescription?: string;
  decimalPoints?: number;
}
export interface BillwiseData {
  accTransactionDetailID: number;
  adjustedAmount: number;
  amount: number;
  balance: number;
  balanceAfter: number;
  billWiseMasterID: number;
  billwiseAmount: number;
  drCr: "Dr" | "Cr"; // Enum-like string values for debit or credit
  financialYearID: number;
  formType: string;
  ledgerID: number;
  partyName: string;
  referenceDate: string; // ISO date string
  referenceNumber: string;
  slNo: number;
  transactionDate: string; // ISO date string
  voucherNumber: number;
  voucherPrefix: string;
  voucherType: string;
}